+++ src/pages/tools.js
@@ -119,24 +119,30 @@ export const Head = () => (
             {
               '@type': 'ListItem',
               position: 6,
+              url: 'https://microlink.io/tools/sitemap',
+              name: 'Sitemap URLs'
+            },
+            {
+              '@type': 'ListItem',
+              position: 7,
               url: 'https://microlink.io/tools/website-to-pdf',
               name: 'Website to PDF'
             },
             {
               '@type': 'ListItem',
-              position: 7,
+              position: 8,
               url: 'https://microlink.io/tools/website-to-pdf/bulk',
               name: 'Bulk PDF'
             },
             {
               '@type': 'ListItem',
-              position: 8,
+              position: 9,
               url: 'https://microlink.io/tools/url-to-markdown',
               name: 'URL to Markdown'
             },
             {
               '@type': 'ListItem',
-              position: 9,
+              position: 10,
               url: 'https://microlink.io/tools/embed-url',
               name: 'Embed URL'
             }